Add basic chain implementation
Allows adding arbitrary data blocks to the chain, and ensuring
validitity. This in itself doesn't mean much until it's deserialisable
and serialisable on a different machine and can be shared, because the
public interfaces ensures it can't directly be changed anyway, but I
wanted to make a chain that could store data and check validity as a
first step.